sql, repr: Add bytes type for sql value (#24252)
<!--
Describe the contents of the PR briefly but completely.

If you write detailed commit messages, it is acceptable to copy/paste
them
here, or write "see commit messages for details." If there is only one
commit
in the PR, GitHub will have already added its commit message above.
-->
For COPY to S3 SQL ([design
doc](https://github.com/moulimukherjee/materialize/blob/design-doc-copy-to-s3/doc/developer/design/20231117_copy_to_s3.md#option-max-file-size))
we'll need a new option type to accept bytes either as a formatted
string like '120 MB' or an integer for the number of bytes.

This PR adds such a type and is also updating existing
`max_query_result_size` to use this type.

From postgres' config settings
https://www.postgresql.org/docs/current/config-setting.html#CONFIG-SETTING-NAMES-VALUES

> Numeric with Unit: Some numeric parameters have an implicit unit,
because they describe quantities of memory or time. The unit might be
bytes, kilobytes, blocks (typically eight kilobytes), milliseconds,
seconds, or minutes. An unadorned numeric value for one of these
settings will use the setting's default unit, which can be learned from
pg_settings.unit. For convenience, settings can be given with a unit
specified explicitly, for example '120 ms' for a time value, and they
will be converted to whatever the parameter's actual unit is. Note that
the value must be written as a string (with quotes) to use this feature.
The unit name is case-sensitive, and there can be whitespace between the
numeric value and the unit.

> Valid memory units are B (bytes), kB (kilobytes), MB (megabytes), GB
(gigabytes), and TB (terabytes). The multiplier for memory units is
1024, not 1000.

> Valid time units are us (microseconds), ms (milliseconds), s
(seconds), min (minutes), h (hours), and d (days).

> If a fractional value is specified with a unit, it will be rounded to
a multiple of the next smaller unit if there is one. For example, 30.1
GB will be converted to 30822 MB not 32319628902 B. If the parameter is
of integer type, a final rounding to integer occurs after any unit
conversion.

### Motivation
Fixes https://github.com/MaterializeInc/materialize/issues/24217

<!--
Which of the following best describes the motivation behind this PR?

  * This PR fixes a recognized bug.

    [Ensure issue is linked somewhere.]

  * This PR adds a known-desirable feature.

    [Ensure issue is linked somewhere.]

  * This PR fixes a previously unreported bug.

    [Describe the bug in detail, as if you were filing a bug report.]

  * This PR adds a feature that has not yet been specified.

[Write a brief specification for the feature, including justification
for its inclusion in Materialize, as if you were writing the original
     feature specification.]

   * This PR refactors existing code.

[Describe what was wrong with the existing code, if it is not obvious.]
-->

### Tips for reviewer
Comments inline

<!--
Leave some tips for your reviewer, like:

    * The diff is much smaller if viewed with whitespace hidden.
    * [Some function/module/file] deserves extra attention.
* [Some function/module/file] is pure code movement and only needs a
skim.

Delete this section if no tips.
-->

### Checklist

- [x] This PR has adequate test coverage / QA involvement has been duly
considered.
- [x] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [x] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [x] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [ ] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
- <!-- Add release notes here or explicitly state that there are no
user-facing behavior changes. -->

+++ src/compute/src/compute_state.rs
@@ -91,7 +91,7 @@ pub struct ComputeState {
     /// History of commands received by this workers and all its peers.
     pub command_history: ComputeCommandHistory<UIntGauge>,
     /// Max size in bytes of any result.
-    max_result_size: u32,
+    max_result_size: u64,
     /// Maximum number of in-flight bytes emitted by persist_sources feeding dataflows.
     pub dataflow_max_inflight_bytes: Option<usize>,
     /// Specification for rendering linear joins.
@@ -127,7 +127,7 @@ impl ComputeState {
             compute_logger: None,
             persist_clients,
             command_history,
-            max_result_size: u32::MAX,
+            max_result_size: u64::MAX,
             dataflow_max_inflight_bytes: None,
             linear_join_spec: Default::default(),
             metrics,
@@ -926,7 +926,7 @@ impl IndexPeek {
     fn seek_fulfillment(
         &mut self,
         upper: &mut Antichain<Timestamp>,
-        max_result_size: u32,
+        max_result_size: u64,
     ) -> Option<PeekResponse> {
         self.trace_bundle.oks_mut().read_upper(upper);
         if upper.less_equal(&self.peek.timestamp) {
@@ -957,7 +957,7 @@ impl IndexPeek {
     /// Collects data for a known-complete peek from the ok stream.
     fn collect_finished_data(
         &mut self,
-        max_result_size: u32,
+        max_result_size: u64,
     ) -> Result<Vec<(Row, NonZeroUsize)>, String> {
         // Check if there exist any errors and, if so, return whatever one we
         // find first.
@@ -989,7 +989,7 @@ impl IndexPeek {
     /// arrangement key-value types.
     fn dispatch_collect_ok_finished_data(
         &mut self,
-        max_result_size: u32,
+        max_result_size: u64,
     ) -> Result<Vec<(Row, NonZeroUsize)>, String> {
         let peek = &mut self.peek;
         let oks = self.trace_bundle.oks_mut();
@@ -1025,7 +1025,7 @@ impl IndexPeek {
         oks_handle: &mut TraceAgent<Tr>,
         key_types: Option<&[ColumnType]>,
         val_types: Option<&[ColumnType]>,
-        max_result_size: u32,
+        max_result_size: u64,
     ) -> Result<Vec<(Row, NonZeroUsize)>, String>
     where
         Tr: TraceReader<Time = Timestamp, Diff = Diff>,
